Fix JSON Totals latency precision and add integrity regression tests (#364)
The Totals section of --json-out-file reported "Latency" and "Average
Latency" as integer-truncated values. A real average of 1.291 ms came
out as 1.000; sub-millisecond averages came out as 0.000. The Sets and
Gets sections were correct because totals_cmd::m_latency is double; only
the outer totals::m_latency was unsigned long long, and run_stats::summarize
assigned the averaged double into it, truncating fractional milliseconds.

Three changes:

- run_stats_types.h: change totals::m_latency from unsigned long long to
  double so the averaged value survives summarize().

- run_stats_types.cpp: fix a long-standing typo in totals::add() that
  conflated m_latency and m_total_latency. The two fields held the same
  value before summarize() ran, so the typo was harmless on the live
  accumulator path but corrupted m_total_latency once summarize() turned
  m_latency into an average for the multi-run aggregate path.

- run_stats.cpp: have summarize() also write the raw latency sum into
  result.m_total_latency, so the BEST/WORST/AGGREGATED sections produced
  by aggregate_average() carry a correct "Accumulated Latency".

Tests: tests/test_json_output_integrity.py asserts the invariant
Latency == Average Latency == Accumulated / Count for Sets, Gets, Totals
across single-run, SET-only, GET-only and multi-run aggregated outputs,
plus the per-second time-series buckets. The existing total_latency > 0
check was insufficient because the bug surfaced as 1.000 ms (looks
plausible) rather than 0.

Refs: RED-191460

+++ run_stats_types.h
@@ -202,7 +202,7 @@ public:
     double m_misses_sec;
     double m_moved_sec;
     double m_ask_sec;
-    unsigned long long int m_latency;
+    double m_latency;
     unsigned long long int m_total_latency;
     // number of bytes received
     unsigned long int m_bytes_rx;
